Выполнил задание 1 из 5 мдуля
