+++ js/id-creator.js
@@ -1,0 +1,7 @@
+const createId = (initialId = 0) => {
+  let currentId = initialId - 1;
+
+  return () => ++currentId;
+};
+
+export { createId };
